# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Assetly API
  version: 1.0.0
  description: API para gerenciamento de ativos e sincronizacao de dados de mercado.
servers:
  - url: /
tags:
  - name: Assets
paths:
  /api/assets:
    get:
      tags: [Assets]
      summary: Lista ativos
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
        - in: query
          name: search
          description: Busca por ticker ou nome do ativo.
          schema:
            type: string
        - in: query
          name: q
          description: Alias para search.
          schema:
            type: string
        - in: query
          name: exchange
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: sector
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, symbol, current_price, updated_at]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Lista paginada de ativos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
    post:
      tags: [Assets]
      summary: Cria um ativo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '201':
          description: Ativo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Simbolo duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/assets/{asset_id}:
    get:
      tags: [Assets]
      summary: Busca um ativo
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Detalhes do ativo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDetails'
        '404':
          description: Ativo nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
    put:
      tags: [Assets]
      summary: Atualiza totalmente um ativo
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '200':
          description: Ativo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ativo nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '409':
          description: Simbolo duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Assets]
      summary: Atualiza parcialmente um ativo
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Ativo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ativo nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '409':
          description: Simbolo duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Assets]
      summary: Remove um ativo
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '204':
          description: Ativo removido
        '404':
          description: Ativo nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
  /api/assets/{asset_id}/sync:
    post:
      tags: [Assets]
      summary: Sincroniza dados de mercado de um ativo
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Sincronizacao concluida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetSyncResponse'
        '404':
          description: Ativo nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '502':
          description: Falha ao consultar provedor de mercado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    AssetId:
      in: path
      name: asset_id
      required: true
      schema:
        type: integer
  schemas:
    AssetInput:
      type: object
      required: [symbol, name, asset_type, currency]
      properties:
        symbol:
          type: string
          example: AAPL
        name:
          type: string
          example: Apple Inc.
        asset_type:
          type: string
          example: equity
        currency:
          type: string
          example: USD
        exchange:
          type: string
          nullable: true
          example: NASDAQ
        sector:
          type: string
          nullable: true
          example: Technology
        current_price:
          type: number
          nullable: true
          example: 170.12
        open_price:
          type: number
          nullable: true
        close_price:
          type: number
          nullable: true
        day_high:
          type: number
          nullable: true
        day_low:
          type: number
          nullable: true
        volume:
          type: integer
          nullable: true
        market_cap:
          type: number
          nullable: true
        pe_ratio:
          type: number
          nullable: true
        dividend_yield:
          type: number
          nullable: true
        external_api_url:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Asset:
      allOf:
        - $ref: '#/components/schemas/AssetInput'
        - type: object
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
              nullable: true
            updated_at:
              type: string
              format: date-time
              nullable: true
    MarketSnapshot:
      type: object
      properties:
        id:
          type: integer
        asset_id:
          type: integer
        price:
          type: number
        open_price:
          type: number
          nullable: true
        close_price:
          type: number
          nullable: true
        high_price:
          type: number
          nullable: true
        low_price:
          type: number
          nullable: true
        volume:
          type: integer
          nullable: true
        captured_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
    SyncLog:
      type: object
      properties:
        id:
          type: integer
        asset_id:
          type: integer
        provider_name:
          type: string
          example: yahoo_finance
        status:
          type: string
          enum: [pending, success, failed]
        message:
          type: string
          nullable: true
        requested_url:
          type: string
          nullable: true
        response_time_ms:
          type: integer
          nullable: true
        synced_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
    AssetDetails:
      allOf:
        - $ref: '#/components/schemas/Asset'
        - type: object
          properties:
            market_snapshots:
              type: array
              items:
                $ref: '#/components/schemas/MarketSnapshot'
            last_sync:
              allOf:
                - $ref: '#/components/schemas/SyncLog'
              nullable: true
    AssetListResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
        offset:
          type: integer
          nullable: true
        limit:
          type: integer
          nullable: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
    AssetSyncResponse:
      type: object
      properties:
        message:
          type: string
          example: sync_completed
        asset_id:
          type: integer
        asset:
          $ref: '#/components/schemas/Asset'
        market_snapshot:
          $ref: '#/components/schemas/MarketSnapshot'
        last_sync:
          $ref: '#/components/schemas/SyncLog'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    NotFoundResponse:
      type: object
      properties:
        message:
          type: string